+++ exercises/11_2DArrays.c
@@ -4,9 +4,9 @@
 int main(int argc, char const *argv[])
 {
   int numbers[][3] = {{1,2,3},{4,5,6}};
-  for (int i = 0; i < sizeof(numbers)/sizeof(numbers[0]); i++)
+  for (size_t i = 0; i < sizeof(numbers)/sizeof(numbers[0]); i++)
   {
-    for (int j = 0; j < sizeof(numbers[0])/sizeof(numbers[0][0]); j++)
+    for (size_t j = 0; j < sizeof(numbers[0])/sizeof(numbers[0][0]); j++)
     {
       printf("%d\n", numbers[i][j]);
     }   
@@ -14,9 +14,9 @@ int main(int argc, char const *argv[])
 
   char cars[][10] = {"Mustang","Corvette","Camaro"};
   strcpy(cars[0], "Bentley");
-  for (int i = 0; i < sizeof(cars)/sizeof(cars[0]); i++)
+  for (size_t i = 0; i < sizeof(cars)/sizeof(cars[0]); i++)
   {
-    printf("car[%d]=%s\n", i, cars[i]);
+    printf("car[%ld]=%s\n", i, cars[i]);
   }
 
   return 0;
